# WebhookForCustomConnector/OpenApiDefinition/OpenApiDefinitionSecure.json
{
  "swagger": "2.0",
  "info": {
    "title": "Exemple de connecteur personnalisé",
    "version": "v1"
  },
  "host": "webhookforcustomconnector.azurewebsites.net",
  "basePath": "/",
  "schemes": [ "https" ],
  "consumes": [ "application/json" ],
  "produces": [ "application/json" ],
  "securityDefinitions": {
    "oauth2_auth": {
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://login.windows.net/common/oauth2/authorize",
      "tokenUrl": "https://login.windows.net/common/oauth2/authorize"
    }
  },
  "paths": {
    "/event/neworder": {
      "x-ms-notification-content": {
        "description": "Création d'une nouvelle commande",
        "schema": {
          "$ref": "#/definitions/Order"
        }
      },
      "post": {
        "description": "Lorsqu'une nouvelle commande est créée (version d'évaluation)",
        "summary": "Lorsqu'une nouvelle commande est créée  (version d'évaluation)",
        "operationId": "NewEcommerceOrder",
        "x-ms-trigger": "single",
        "parameters": [
          {
            "in": "body",
            "name": "Callback",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Callback"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/event/instore": {
      "x-ms-notification-content": {
        "description": "Arrivée de nouveaux produits",
        "schema": {
          "$ref": "#/definitions/InStore"
        }
      },
      "post": {
        "description": "Lorsque qu'un nouveau produit arrive dans le magasin (version d'évaluation)",
        "summary": "Lorsque qu'un nouveau produit arrive dans le magasin (version d'évaluation)",
        "operationId": "NewInstoreProduct",
        "x-ms-trigger": "single",
        "parameters": [
          {
            "in": "body",
            "name": "Callback",
            "schema": {
              "$ref": "#/definitions/Callback"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        }
      }
    },
    "/event/remove/{oid}/{id}": {
      "delete": {
        "description": "Supprimer l'abonnement",
        "summary": "Supprimer l'abonnement",
        "operationId": "RemoveSubscription",
        "x-ms-visibility": "internal",
        "parameters": [
          {
            "in": "path",
            "name": "oid",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identification de l'abonnement",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/list/InStore": {
      "get": {
        "description": "Récupère la liste des nouveaux produits en magasin (version d'évaluation)",
        "summary": "Récupère la liste des nouveaux produits en magasin (version d'évaluation)",
        "operationId": "GetListInstore",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/InStore"
              }
            }
          }
        }
      }
    },
    "/list/neworders": {
      "get": {
        "description": "Récupère la liste des nouvelles commandes (version d'évaluation)",
        "summary": "Récupère la liste des nouvelles commandes (version d'évaluation)",
        "operationId": "GetListNewOrders",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Order"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Callback": {
      "type": "object",
      "required": [ "Url" ],
      "properties": {
        "Url": {
          "x-ms-notification-url": true,
          "x-ms-visibility": "internal",
          "required": [ "Url" ],
          "description": "URL de rappel",
          "title": "URL de rappel",
          "type": "string"
        }
      }
    },
    "InStore": {
      "type": "object",
      "properties": {
        "storeName": {
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "OrderDetails": {
      "type": "object",
      "properties": {
        "orders": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {        
        "customerName": {
          "type": "string"
        },
        "orderNumber": {
          "type": "string"
        },
        "total": {
          "format": "double",
          "type": "number"
        }
      }
    }
  }
}